+++ .cicd/commands/main.sh
@@ -4,32 +4,36 @@ export CARLIN_ENVIRONMENT=Production
 git fetch --tags --quiet
 
 # https://stackoverflow.com/a/25742085/8786986
-yarn lerna changed || { echo "No changes detected, exiting main workflow" && exit 0; }
+pnpm lerna changed || { echo "No changes detected, exiting main workflow" && exit 0; }
 
-####
-## If we're here, there are changes, so we need to run the main workflow
-####
+# If we're here, there are changes, so we need to run the main workflow
 
 LATEST_TAG=$(git describe --tags --abbrev=0)
 
+# Used because of https://github.com/vercel/turbo/issues/4559 issue.
+LATEST_TAG_SHA=$(git rev-list -n 1 $LATEST_TAG)
+
 # Setup NPM token
-echo //registry.npmjs.org/:\_authToken=$NPM_TOKEN > .npmrc
+# Using ~/.npmrc instead of .npmrc because pnpm uses .npmrc and appending
+# the token to .npmrc will cause git uncommitted changes error.
+echo //registry.npmjs.org/:\_authToken=$NPM_TOKEN > ~/.npmrc
 
 # Print "NPM whoami" to check if the token is valid
 echo NPM whoami: $(npm whoami)
 
-# Lint
-yarn turbo run lint
-
-# Build config to run lint-staged for lint and version bump
-yarn turbo run build --filter=@ttoss/config...
+# Build @ttoss/config package to lerna version command works properly
+# when commiting changes. If we don't build this package, commit will fail
+# because pre-commit hook will run syncpack:list with default config, that
+# not works because of package version and "workspace:^" mismatch.
+pnpm turbo run build:config
 
 # Version before publish to rebuild all packages that Lerna will publish
-yarn lerna version --yes --no-push
+pnpm lerna version --yes --no-push
 
-# Lint, test, and build all packages since $LATEST_TAG and their dependent packages.
+# Test and build all packages since $LATEST_TAG_SHA
+# and all the workspaces that depends on them
 # https://turbo.build/repo/docs/core-concepts/monorepos/filtering#include-dependents-of-matched-workspaces
-yarn turbo run build test --filter=...[$LATEST_TAG]
+pnpm turbo run build test --filter=...[$LATEST_TAG_SHA]
 
 # Undo all files that were changed by the build command—this happens because
 # the build can change files with different linting rules, or modify some
@@ -38,7 +42,7 @@ yarn turbo run build test --filter=...[$LATEST_TAG]
 git checkout -- .
 
 # Publish packages
-yarn lerna publish from-git --yes
+pnpm lerna publish from-git --yes
 
 # Push only tags to check if there's no issues with the tags
 git push --tags
@@ -49,5 +53,5 @@ git push --follow-tags
 # Deploy after publish because there are cases in which a package is versioned
 # and it should be on NPM registry to Lambda Layer create the new version when
 # carlin deploy starts.
-yarn turbo run deploy --filter=...[$LATEST_TAG]
+pnpm turbo run deploy --filter=...[$LATEST_TAG_SHA]