+++ .cicd/commands/pr.sh
@@ -1,9 +1,4 @@
-# Lint
-yarn turbo run lint
-
-# Build config to run lint-staged for lint and version bump
-yarn turbo run build --filter=@ttoss/config...
-
-# Test, build, and deploy all packages since main and their dependent packages.
+# Test, build, and deploy all packages since main
+# and all the workspaces that depends on them.
 # https://turbo.build/repo/docs/core-concepts/monorepos/filtering#include-dependents-of-matched-workspaces
-yarn turbo run build test deploy --filter=...[main]
+pnpm turbo run build test deploy --filter=...[main]