+++ docs/storybook/stories/forms/FormGroup.stories.tsx
@@ -1,6 +1,6 @@
 import { Button } from '@ttoss/ui';
 import { Form, FormFieldInput, FormGroup } from '@ttoss/forms/src';
-import { Meta, Story } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { useForm } from 'react-hook-form';
 
@@ -9,7 +9,7 @@ export default {
   component: FormGroup,
 } as Meta;
 
-const Template: Story = () => {
+const Template: StoryFn = () => {
   const formMethods = useForm({
     mode: 'all',
   });
@@ -24,29 +24,46 @@ const Template: Story = () => {
           <FormFieldInput name="input3" placeholder="input3" label="input3" />
           <FormFieldInput name="input4" placeholder="input4" label="input4" />
 
-          <FormGroup>
+          <FormGroup direction="row">
             <FormFieldInput name="input5" placeholder="input5" label="input5" />
 
-            <FormGroup>
+            <FormGroup direction="row">
               <FormFieldInput
-                name="input6"
-                placeholder="input6"
-                label="input6"
+                name="input6-1"
+                placeholder="input6-1"
+                label="input6-1"
               />
-              <FormGroup>
-                <FormFieldInput
-                  name="input9"
-                  placeholder="input9"
-                  label="input9"
-                />
-              </FormGroup>
+              <FormFieldInput
+                name="input7-1"
+                placeholder="input7-1"
+                label="input7-1"
+              />
+            </FormGroup>
+
+            <FormGroup direction="row">
+              <FormFieldInput
+                name="input6-2"
+                placeholder="input6-2"
+                label="input6-2"
+              />
+              <FormFieldInput
+                name="input7-2"
+                placeholder="input7-2"
+                label="input7-2"
+              />
             </FormGroup>
           </FormGroup>
 
-          <FormGroup>
-            <FormFieldInput name="input7" placeholder="input7" label="input7" />
+          <FormGroup direction="row">
             <FormFieldInput name="input8" placeholder="input8" label="input8" />
+            <FormFieldInput name="input9" placeholder="input9" label="input9" />
           </FormGroup>
+
+          <FormFieldInput
+            name="input10"
+            placeholder="input10"
+            label="input10"
+          />
         </FormGroup>
       </FormGroup>