+++ docs/storybook/stories/forms/Radio.stories.tsx
@@ -1,13 +1,13 @@
 import * as yup from 'yup';
 import { Button } from '@ttoss/ui';
 import { Form, FormFieldRadio } from '@ttoss/forms';
-import { Meta, Story } from '@storybook/react';
+import { Meta, StoryFn } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 
 export default {
-  title: 'Forms/FormField.Radio',
+  title: 'Forms/FormFieldRadio',
   component: FormFieldRadio,
 } as Meta;
 
@@ -15,7 +15,7 @@ const schema = yup.object({
   car: yup.string().required('Car is required'),
 });
 
-const Template: Story = () => {
+const Template: StoryFn = () => {
   const formMethods = useForm({
     mode: 'all',
     resolver: yupResolver(schema),