+++ packages/ui/src/index.ts
@@ -35,3 +35,4 @@ export { Container, type ContainerProps } from './components/Container';
 export { HelpText, type HelpTextProps } from './components/HelpText';
 export { CloseButton, type CloseButtonProps } from './components/CloseButton';
 export { InputNumber, type InputNumberProps } from './components/InputNumber';
+export { Stack, type StackProps } from './components/Stack';