+++ docs/storybook/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.35.0](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.34.11...@docs/storybook@1.35.0) (2023-04-14)
+
+### Features
+
+- add stack component ([#271](https://github.com/ttoss/ttoss/issues/271)) ([275c559](https://github.com/ttoss/ttoss/commit/275c5593edaa5f3b29e40d7f349c6a1bd8e886e1))
+
 ## [1.34.11](https://github.com/ttoss/ttoss/compare/@docs/storybook@1.34.10...@docs/storybook@1.34.11) (2023-04-13)
 
 **Note:** Version bump only for package @docs/storybook