+++ packages/test-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.21.3 (2023-04-16)
+
+**Note:** Version bump only for package @ttoss/test-utils
+
 ## [1.21.2](https://github.com/ttoss/ttoss/compare/@ttoss/test-utils@1.21.1...@ttoss/test-utils@1.21.2) (2023-04-05)
 
 **Note:** Version bump only for package @ttoss/test-utils